Separate and update Geo Location configuration
