+++ docs/config-app.md
@@ -229,7 +229,10 @@ If not defined in config all other Health Checkers would be disabled and endpoin
 - `gdpr.vendorlist.http-endpoint-template` - template string for vendor list url, where `{VERSION}` is used as version number placeholder.
 - `gdpr.vendorlist.http-default-timeout-ms` - default operation timeout for obtaining new vendor list.
 - `gdpr.vendorlist.filesystem-cache-dir` - directory for local storage cache for vendor list. Should be with `WRITE` permissions for user application run from.
-- `gdpr.geolocation.enabled` - if equals to `true` the geo location service will be used to determine the country for client request.
+
+## Geo Location
+- `geolocation.enabled` - if equals to `true` the geo location service will be used to determine the country for client request.
+- `geolocation.type` - name of the geo location implementation provider that should be used.
 
 ## Auction (Legacy)
 - `default-timeout-ms` - this setting controls default timeout for /auction endpoint.
